+++ Source/BoomBoxPlus/Private/Tape/BBPCustomMusicTape.cpp
@@ -5,7 +5,7 @@
 UBBPCustomMusicTape::UBBPCustomMusicTape()
 {
 	mTitle = LOCTEXT("CustomMusicTitle", "Custom Music");
-	mDescription = LOCTEXT("CustomMusicDescription", "Your own music, YouTube and SoundCloud.");
+	mDescription = GetIdleDescription();
 
 	const TSoftObjectPtr<UTexture2D> EmptyTapeIcon(FSoftObjectPath(TEXT("/Game/FactoryGame/Equipment/BoomBox/TXUI_Tape_EmptyTape.TXUI_Tape_EmptyTape")));
 	mSmallIcon = EmptyTapeIcon;
@@ -15,6 +15,11 @@ UBBPCustomMusicTape::UBBPCustomMusicTape()
 bool UBBPCustomMusicTape::IsCustomMusicTape(TSubclassOf<UFGTapeData> Tape)
 {
 	return Tape && Tape->IsChildOf(UBBPCustomMusicTape::StaticClass());
+}
+
+FText UBBPCustomMusicTape::GetIdleDescription()
+{
+	return LOCTEXT("CustomMusicDescription", "Your own music, YouTube and SoundCloud.");
 }
 
 #undef LOCTEXT_NAMESPACE